fix(scout-for-lol): count emoji display widths
Why
Codex review found that variation-selector emoji such as heart and copyright can render at two Discord columns while the table formatter counted one.

What
Centralize native table display-width calculation with the formatting helpers, count emoji-presentation variation selectors alongside flags and keycaps, and cover all cases with direct width regression tests.

Verification
- bun test src/discord/scout/visualization.test.ts
- bunx eslint src/discord/scout/visualization.ts src/discord/scout/visualization-format.ts src/discord/scout/visualization.test.ts --quiet
- bun run typecheck
- bunx lefthook run pre-commit
